+++ job_radar/company_catalog_query_service.py
@@ -28,6 +28,7 @@ _URL_SOURCE_TYPES = {
     "jobsyn",
     "oracle_hcm",
     "phenom",
+    "recruitee",
     "selectminds",
     "smartrecruiters",
     "weka",